requirements.txt
# YouTube Prosperity Sermon Corpus Builder
# See build_corpus.py docstring for setup instructions

google-api-python-client
youtube-transcript-api>=1.0
pandas
python-dateutil
tqdm
yt-dlp

# LLM extraction (run_extraction.py)
google-genai

# Dashboard & NLP analysis
streamlit>=1.28
plotly>=5.18
scikit-learn>=1.3
umap-learn>=0.5
numpy>=1.24
